+++ package.json
@@ -23,8 +23,5 @@
     "lerna": "^8.1.2",
     "lint-staged": "^15.2.2",
     "prettier": "^3.2.5"
-  },
-  "dependencies": {
-    "ethernaut-util": "^1.1.11"
   }
 }
